로그인 기능 구현 업로드

+++ Chapter06/src/main/java/edu/pnu/controller/BoardController.java
@@ -4,17 +4,26 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.SessionAttributes;
 
 import edu.pnu.domain.Board;
+import edu.pnu.domain.Member;
 import edu.pnu.service.BoardServiceImpl;
 
+@SessionAttributes("member")
 @Controller
 public class BoardController {
 
 	@Autowired
 	private BoardServiceImpl boardService;
+
+	@ModelAttribute("member")
+	public Member setMember() {
+		return new Member();
+	}
 
 	@GetMapping("/hello")
 	public void hello(Model model) {
@@ -22,35 +31,55 @@ public class BoardController {
 	}
 
 	@RequestMapping("/getBoardList")
-	public String getBoardList(Model model) {
+	public String getBoardList(@ModelAttribute("member") Member member, Model model, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
 		model.addAttribute("boardList", boardService.getBoardList());
 		return "getBoardList";
 	}
 
 	@GetMapping("/insertBoard")
-	public void insertBoard() {
+	public String insertBoard(@ModelAttribute("member") Member member) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
+
+		return "insertBoard";
 	}
 
 	@PostMapping("/insertBoard")
-	public String insertBoard(Board board) {
+	public String insertBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
 		boardService.insertBoard(board);
 		return "redirect:getBoardList";
 	}
 
 	@GetMapping("/getBoard")
-	public String getBoard(Board board, Model model) {
+	public String getBoard(@ModelAttribute("member") Member member, Board board, Model model) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
 		model.addAttribute("board", boardService.getBoard(board));
 		return "getBoard";
 	}
 
 	@PostMapping("/updateBoard")
-	public String updateBoard(Board board) {
+	public String updateBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
 		boardService.updateBoard(board);
 		return "forward:getBoardList";
 	}
 
 	@GetMapping("/deleteBoard")
-	public String deleteBoard(Board board) {
+	public String deleteBoard(@ModelAttribute("member") Member member, Board board) {
+		if (member.getId() == null) {
+			return "redirect:login";
+		}
 		boardService.deleteBoard(board);
 		return "forward:getBoardList";
 	}
